TODO: Implement workaround to MapKit bug
According to my research the crashes are the result of a bug in MapKit,
wherein an exception is thrown in Objc when MK attepmts to update
annotations while user interaction is taking place. In my case, this
occurs when the map is being panned by the user while MK attempts to
recalculate clusters. The cluster recalculation is driven by zooming and
adding or removing annotations. A possible workaround is to create some
sort of mutex on annotation insertions controlled by
mapViewRegionDidChange and mapViewRegionWillChange